IEndpointFactory interception point added.
